test(processes): merge processes_review2_test.go into processes_test.go
